Fix some reset code; Add state flags; Add some verbose messages
Fixed resetChannelStates.  This was commented out and just had a BRK an
the entrypoint.  The states had been reorganized in RAM so this was
disabled.

Added a verbose listing of all the channel state variables that are
generated for each channel.

Added engine state flags for Ready, EnableSong, and EnableSfx.  If Ready
isn't set when  SoundProcess is called, SoundInit is called to setup the
engine's initial state.  If EnableSong or EnableSfx are aren't set then
the song or Sfx won't play (respectively).  They should still keep their
state, however.