Fix song loading; Fix tick values
Fix loading the song so it actually uses the argument passed in A as the
index into the song list, rather than the offset into the song's
metadata (screwing up everything).

Properly load the song's Speed value into Global::Ticks when appropriate
and initialize Global::Ticks to 1 so the engine is updated on the next
process cycle.