Implement noise channel
Removed the noise-specific routine to parse the noise channel and just
use the same one that all the other channels are using.  Added a check
on the channel ID to differentiate between noise and not noise when it
comes to note lookups.

Also, write a static non-zero value to all the channels' Length Counter
registers.